chore: fix domain package
