+++ src/main/kotlin/com/example/anbdapi/domain/auth/dto/request/LogoutRequest.kt
@@ -1,0 +1,5 @@
+package com.example.anbdapi.domain.auth.dto.request
+
+data class LogoutRequest(
+    val email: String
+)
