+++ src/main/kotlin/com/example/anbdapi/domain/auth/dto/request/RefreshRequest.kt
@@ -1,0 +1,7 @@
+package com.example.anbdapi.domain.auth.dto.request
+
+data class RefreshRequest(
+    val email: String,
+    val accessToken: String,
+    val refreshToken: String
+)
