+++ src/main/kotlin/com/example/anbdapi/domain/auth/dto/response/LoginResponse.kt
@@ -1,0 +1,7 @@
+package com.example.anbdapi.domain.auth.dto.response
+
+data class LoginResponse(
+    val accessToken: String,
+    val refreshToken: String,
+    val profileComplete: Int
+)
