+++ src/main/kotlin/com/example/anbdapi/domain/auth/dto/response/TokenResponse.kt
@@ -1,0 +1,6 @@
+package com.example.anbdapi.domain.auth.dto.response
+
+data class TokenResponse(
+    val accessToken: String,
+    val refreshToken: String
+)
