+++ src/main/kotlin/com/example/anbdapi/domain/auth/entity/Gender.kt
@@ -1,0 +1,5 @@
+package com.example.anbdapi.domain.auth.entity
+
+enum class Gender {
+    MALE, FEMALE, OTHER
+}
