+++ src/main/kotlin/com/example/anbdapi/domain/auth/entity/Provider.kt
@@ -1,0 +1,5 @@
+package com.example.anbdapi.domain.auth.entity
+
+enum class Provider {
+        GOOGLE, APPLE
+}
