+++ src/main/kotlin/com/example/anbdapi/domain/auth/entity/ShareCategory.kt
@@ -1,0 +1,5 @@
+package com.example.anbdapi.domain.auth.entity
+
+enum class ShareCategory {
+    FOOD, DIGITAL
+}
