+++ src/main/kotlin/com/example/anbdapi/domain/auth/exception/TokenExpiredException.kt
@@ -1,0 +1,3 @@
+package com.example.anbdapi.domain.auth.exception
+
+class TokenExpiredException(message: String) : RuntimeException(message)
