+++ src/main/kotlin/com/example/anbdapi/domain/auth/service/AuthService.kt
@@ -1,0 +1,5 @@
+package com.example.anbdapi.domain.auth.service
+
+interface AuthService {
+    fun refreshAccessToken(request: com.example.anbdapi.domain.auth.dto.request.RefreshRequest): com.example.anbdapi.domain.auth.dto.response.TokenResponse
+}
