--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/dto/request/LogoutRequest.kt
@@ -1,5 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.dto.request
-
-data class LogoutRequest(
-    val email: String
-)
