--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/dto/request/RefreshRequest.kt
@@ -1,7 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.dto.request
-
-data class RefreshRequest(
-    val email: String,
-    val accessToken: String,
-    val refreshToken: String
-)
