--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/dto/response/LoginResponse.kt
@@ -1,7 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.dto.response
-
-data class LoginResponse(
-    val accessToken: String,
-    val refreshToken: String,
-    val profileComplete: Int
-)
