--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/dto/response/TokenResponse.kt
@@ -1,6 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.dto.response
-
-data class TokenResponse(
-    val accessToken: String,
-    val refreshToken: String
-)
