--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/entity/Gender.kt
@@ -1,5 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.entity
-
-enum class Gender {
-    MALE, FEMALE, OTHER
-}
