--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/entity/ShareCategory.kt
@@ -1,5 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.entity
-
-enum class ShareCategory {
-    FOOD, DIGITAL
-}
