--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/exception/TokenExpiredException.kt
@@ -1,3 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.exception
-
-class TokenExpiredException(message: String) : RuntimeException(message)
