--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/repository/UserRepository.kt
@@ -1,11 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.repository
-
-import com.example.anbdapi.domain.dev.auth.entity.User
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-import java.util.*
-
-@Repository
-interface UserRepository : JpaRepository<User, Long> {
-    fun findByEmail(email: String): User?
-}
