--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/repository/UserSocialAccountRepository.kt
@@ -1,11 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.repository
-
-import com.example.anbdapi.domain.dev.auth.entity.Provider
-import com.example.anbdapi.domain.dev.auth.entity.UserSocialAccount
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface UserSocialAccountRepository : JpaRepository<UserSocialAccount, Long> {
-    fun findByProviderAndProviderId(provider: Provider, providerId: String): UserSocialAccount?
-}
