--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/service/AuthService.kt
@@ -1,8 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.service
-
-import com.example.anbdapi.domain.dev.auth.dto.request.RefreshRequest
-import com.example.anbdapi.domain.dev.auth.dto.response.TokenResponse
-
-interface AuthService {
-    fun refreshAccessToken(request: RefreshRequest): TokenResponse
-}
