--- src/main/kotlin/com/example/anbdapi/domain/dev/auth/service/CustomOAuth2UserService.kt
@@ -1,8 +1,0 @@
-package com.example.anbdapi.domain.dev.auth.service
-
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest
-import org.springframework.security.oauth2.core.user.OAuth2User
-
-interface CustomOAuth2UserService {
-    fun loadUser(userRequest: OAuth2UserRequest): OAuth2User
-}
